perf(tests): share one Postgres across backend/store
`go test ./backend/store/` took 382.95s, because 90 of its tests each started
their own Postgres container at 4.3s a piece. Give the package a TestMain that
starts one container and migrates a template database once; each test then
copies that template, which costs 44ms.

Same tests, same machine: 382.95s -> 12.0s.

One helper covers every call site. newTestDB returns a raw handle for seeding, a
Store for the code under test, and the URL for the few tests that open their own
connections, all on the test's own database. The copies are not dropped: they
live in the package's own container, which TestMain terminates on the way out,
and dropping each one costs 54ms to reclaim disk that is already reclaimed.

AGENTS.md confines a real metadata Postgres to backend/store and backend/tests,
so ten tests that assert SQL behaviour move out of backend/api/v1 -- the two
audit-log retention tests and the eight issue concurrency, transaction and
pagination tests. Assertions are unchanged. Two things could not follow:
convertToAuditLogs is unexported, so its name-formatting assertion stays behind
as a pure converter test with no database; and the tests asserted on
IssueService.bus, an unexported field, so the helper now returns the bus it
builds.

Nine fixture helpers are duplicated between the two packages for now. Both
copies carry a comment saying so; the duplication ends when the 27 issue tests
still in backend/api/v1 move to a fake.
