+++ backend/store/issue_comment_thread_test.go
@@ -2,15 +2,12 @@ package store_test
 
 import (
 	"context"
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 
 	"github.com/bytebase/bytebase/backend/common"
-	"github.com/bytebase/bytebase/backend/common/testcontainer"
 	storepb "github.com/bytebase/bytebase/backend/generated-go/store"
-	"github.com/bytebase/bytebase/backend/migrator"
 	"github.com/bytebase/bytebase/backend/store"
 
 	_ "github.com/bytebase/bytebase/backend/plugin/db/pg"
@@ -30,10 +27,7 @@ func TestIssueCommentThreads(t *testing.T) {
 	)
 
 	ctx := context.Background()
-	container := testcontainer.GetTestPgContainer(ctx, t)
-	t.Cleanup(func() { container.Close(ctx) })
-	db := container.GetDB()
-	require.NoError(t, migrator.MigrateSchema(ctx, db))
+	db, stores, _ := newTestDB(t)
 
 	_, err := db.ExecContext(ctx,
 		`INSERT INTO workspace (resource_id) VALUES ($1)`, workspaceID)
@@ -51,14 +45,6 @@ func TestIssueCommentThreads(t *testing.T) {
 			id, projectID, creator)
 		require.NoError(t, err)
 	}
-
-	pgURL := fmt.Sprintf(
-		"host=%s port=%s user=postgres password=root-password database=postgres",
-		container.GetHost(), container.GetPort(),
-	)
-	stores, err := store.New(ctx, pgURL, false)
-	require.NoError(t, err)
-	t.Cleanup(func() { require.NoError(t, stores.Close()) })
 
 	anchor := &storepb.IssueCommentPayload_StatementAnchor{
 		SpecId:        "spec-1",
